requirements.txt
# Screen capture and OCR dependencies
mss>=9.0.0              # Fast, cross-platform screenshots
easyocr>=1.7.0          # High-quality OCR with GPU support
httpx>=0.27.0           # Modern async HTTP client
Pillow>=10.0.0          # Image processing
python-dotenv>=1.0.0    # Environment configuration
